Wertebereich in compress_pattern durch cast auf long erhöht

+++ src/_micro-api/libraries/signalDecoder/src/signalDecoder.cpp
@@ -257,7 +257,7 @@ void SignalDetectorClass::compress_pattern()
 		{
 			if (histo[idx2] == 0 || (pattern[idx] ^ pattern[idx2]) >> 15)
 				continue;
-			const int16_t tol = int(((abs(pattern[idx2])*tolFact) + (abs(pattern[idx])*tolFact)) / 2);
+			const int16_t tol = int(((abs(long(pattern[idx2]))*tolFact) + (abs(pattern[idx])*tolFact)) / 2);
 #if DEBUGDETECT>2
 			DBG_PRINT("comptol: "); DBG_PRINT(tol); DBG_PRINT("  "); DBG_PRINT(idx2); DBG_PRINT("<->"); DBG_PRINT(idx); DBG_PRINT(";");
 #endif // DEBUGDETECT
